+++ samples/QueueSamples.php
@@ -292,7 +292,7 @@ function createQueueAccountSASSample()
         'rwdlacup',                // Read, Write, Delete, List, Add, Create, Update, Process
         'q',                       // Queue
         'sco',                     // Service, container and object level resources
-        '2020-01-01T08:30:00Z',    // A valid ISO 8601 time format
+        '2030-01-01T08:30:00Z',    // A valid ISO 8601 time format
         '2016-01-01T12:00:00Z',    // A valid ISO 8601 time format
         '1.1.1.1-255.255.255.255', // An IP or IP ranges
         'https,http'               // Protocol permitted for requests